[Windows services]-<File Monitoring Project (In Progress 90%Done)>

+++ 01_Back-end/Part-2-with_C#/09_WindowsServices/wsFileMonitoring_Project/wsFileMonitoring_Project/FileMonitoringService.cs
@@ -1,13 +1,16 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Configuration;
 using System.Data;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
 using System.ServiceProcess;
-using System.Configuration;
+using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
-using System.IO;
 
 namespace wsFileMonitoring_Project
 {
@@ -18,16 +21,19 @@ namespace wsFileMonitoring_Project
         private string _DestinationDirectory;
         private string _LogDirectory;
 
-        private string _TargetFile;
-
         private string _LogFileName;
         private string _LogFilePath;
+
+        private FileSystemWatcher _FileSystemWatcher;
+        readonly ConcurrentQueue<string> TargetFiles = new ConcurrentQueue<string>();
+        private readonly AutoResetEvent _signal = new AutoResetEvent(false);
         public FileMonitoringService()
         {
             InitializeComponent();
 
             CanPauseAndContinue = true;
             CanStop = true;
+            
 
 
             InitializeRequiredDirectories();
@@ -49,7 +55,7 @@ namespace wsFileMonitoring_Project
             _CreateDirectoryIfNoExists(_SourceDirectory);
             _CreateDirectoryIfNoExists(_DestinationDirectory);
             _CreateDirectoryIfNoExists(_LogDirectory);
-
+            _LogFileName = ConfigurationManager.AppSettings["LogFileName"];
         }
         private void _CreateDirectoryIfNoExists(string TargetDirectory)
         {
@@ -63,42 +69,24 @@ namespace wsFileMonitoring_Project
             string logMessage = $"[{DateTime.Now:G}] {message}\n";
             File.AppendAllText(_LogFilePath, logMessage);
         }
-        private void _ConvertFileNameToDUID(string FilePath)
+        private string _RenameFileToGuid(string filePath)
         {
-            try
-            {
-                if (string.IsNullOrWhiteSpace(FilePath))
-                    throw new ArgumentException("File name cannot be null or empty.", nameof(FilePath));
-
-                string extension = Path.GetExtension(FilePath);
-                string guid = Guid.NewGuid().ToString("N");
-
-                string NewName = Path.Combine(Directory.GetParent(FilePath).FullName, $"{guid}{extension}");
-                File.Move(FilePath , NewName);
-            }
-            catch (Exception e)
-            {
-                _LogAction($"Error: {e.Message}");
-            }
             
-        }
-        private string RenameFileToGuid(string filePath)
-        {
-            if (string.IsNullOrWhiteSpace(filePath))
-                throw new ArgumentException("File path cannot be null or empty.", nameof(filePath));
-
-            if (!File.Exists(filePath))
-                throw new FileNotFoundException("File does not exist.", filePath);
 
             try
             {
+                if (string.IsNullOrWhiteSpace(filePath))
+                    throw new ArgumentException("File path cannot be null or empty.", nameof(filePath));
+
+                if (!File.Exists(filePath))
+                    throw new FileNotFoundException("File does not exist.", filePath);
                 string directory = Path.GetDirectoryName(filePath);
                 string extension = Path.GetExtension(filePath);
                 string newFileName = $"{Guid.NewGuid():N}{extension}";
                 string newPath = Path.Combine(directory, newFileName);
 
                 File.Move(filePath, newPath);
-
+                _LogAction($"Rename File To GUID From[{filePath}] =To=> [{newPath}]");
                 return newPath;
             }
             catch (IOException ex)
@@ -107,23 +95,105 @@ namespace wsFileMonitoring_Project
                 throw;
             }
         }
+        private string _ChangeFilePath(string OldfilePath,string DestinationDirectory)
+        {
+
+            string FileName = Path.GetFileName(OldfilePath);
+            string NewPath = Path.Combine(DestinationDirectory, FileName);
+            return NewPath;
+        }
         private void _ChangeFileLocation(string sourcePath, string DestinationPath)
         {
 
             File.Move(sourcePath, DestinationPath);
-            File.Delete(sourcePath);
-            _LogAction($"< Target File Moved > [ {sourcePath} ==> {DestinationPath}");
+            if(File.Exists(sourcePath)) File.Delete(sourcePath); ;
         }
-       
-        
+        private void _WaitForFile(string path)
+        {
+            for (int i = 0; i < 10; i++)
+            {
+                try
+                {
+                    using (FileStream stream = File.Open(path, FileMode.Open, FileAccess.Read, FileShare.None))
+                    {
+                        return;
+                    }
+                }
+                catch
+                {
+                    Task.Delay(500).Wait();
+                }
+            }
+        }
+        private void OnFileInserted(object sender, FileSystemEventArgs e)
+        {
+
+            TargetFiles.Enqueue(e.FullPath);
+            _signal.Set();
+        }
+        private void StartWorker()
+        {
+            Task.Run
+                (
+                    () =>
+                    {
+                        while (true)
+                        {
+                            _signal.WaitOne();
+
+
+                            StringBuilder CurrentFilePath = new StringBuilder();
+                            StringBuilder FilePath_GUID = new StringBuilder();
+                            StringBuilder FileDestinationPath = new StringBuilder();
+
+                            while (TargetFiles.TryDequeue(out var currentFilePath))
+                            {
+
+                                try
+                                {
+                                    CurrentFilePath.Append(CurrentFilePath);
+                                    _WaitForFile(CurrentFilePath.ToString());
+
+                                    FilePath_GUID.Append(_RenameFileToGuid(CurrentFilePath.ToString()));
+                                    FileDestinationPath.Append(_ChangeFilePath(FilePath_GUID.ToString(), _DestinationDirectory));
+
+                                    _ChangeFileLocation(FilePath_GUID.ToString(), FileDestinationPath.ToString());
+
+                                    _LogAction($"< Target File Moved > [ {CurrentFilePath} ==> {FileDestinationPath}");
+                                }
+                                catch (Exception ex)
+                                {
+                                    _LogAction($"Error processing file {CurrentFilePath}: {ex.Message}");
+                                }
+                            }
+                        }
+                    }
+                );
+        }
+
         protected override void OnStart(string[] args)
         {
+            Process process = Process.GetCurrentProcess();
+            process.PriorityClass = ProcessPriorityClass.BelowNormal;
             _LogAction("Service Started");
 
-        }
+            _FileSystemWatcher = new FileSystemWatcher
+            {
+                Path = _MainDirectory,
+                Filter = "*.*",
+                NotifyFilter = NotifyFilters.FileName | NotifyFilters.CreationTime
+            };
 
+            _FileSystemWatcher.Created += OnFileInserted;
+            _FileSystemWatcher.EnableRaisingEvents = true;
+            StartWorker();
+
+        }
+      
         protected override void OnStop()
         {
+            _FileSystemWatcher.EnableRaisingEvents = false;
+            _FileSystemWatcher.Dispose();
             _LogAction("Service Stopped");
         }
